+++ Sources/Swiftor/Vector2.swift
@@ -1,13 +1,13 @@
 public struct Vector2: Vector, Codable {
-    public var x: Float
-    public var y: Float
+    public var x: Double
+    public var y: Double
     
-    public init(x: Float, y: Float) {
+    public init(x: Double, y: Double) {
         self.x = x
         self.y = y
     }
     
-    public func dotProduct(with v: Vector2) -> Float {
+    public func dotProduct(with v: Vector2) -> Double {
         return x * v.x + y * v.y
     }
     
@@ -30,11 +30,11 @@ public func -(lhs: Vector2, rhs: Vector2) -> Vector2 {
     return Vector2(x: lhs.x - rhs.x, y: lhs.y - rhs.y)
 }
 
-public func *(lhs: Vector2, rhs: Float) -> Vector2 {
+public func *(lhs: Vector2, rhs: Double) -> Vector2 {
     return Vector2(x: lhs.x * rhs, y: lhs.y * rhs)
 }
 
-public func /(lhs: Vector2, rhs: Float) -> Vector2 {
+public func /(lhs: Vector2, rhs: Double) -> Vector2 {
     return Vector2(x: lhs.x / rhs, y: lhs.y / rhs)
 }
 
@@ -43,7 +43,7 @@ public func ==(lhs: Vector2, rhs: Vector2) -> Bool {
 }
 
 extension Vector2: ExpressibleByArrayLiteral {
-    public init(arrayLiteral elements: Float...) {
+    public init(arrayLiteral elements: Double...) {
         precondition(elements.count == 2)
         
         self.init(x: elements[0], y: elements[1])
